+++ foam/namespace.py
@@ -1,0 +1,11 @@
+__all__ = ['Command', 'Information', 'PostProcess', 'VTK', 'Foam', 'Data', 'Envelope', 'SMTP', 'Timer', 'Version']
+
+
+from .app.command.core import Command
+from .app.information.core import Information
+from .app.postprocess.core import PostProcess, VTK
+from .base.core import Foam
+from .util.object.data import Data
+from .util.object.email import Envelope, SMTP
+from .util.object.timer import Timer
+from .util.object.version import Version