+++ CHANGELOG.md
@@ -83,6 +83,7 @@
 ## `v0.12`
 
 - Add extra sub-module for sending emails and timing
+- Add namespace to package
 - Update program architecture
     - `foam::base::type::{Data, Version}` -> `foam::util::object::{Data, Version}`
     - `foam::base::parse` -> `foam::parse`